Separated the monolithic interactAgentHomeWork() into separate interaction classes for various locations (home, work, school, neighborhood by creating interaction location classes
