+++ src/InteractionLocations.H
@@ -1,0 +1,13 @@
+/*! @file InteractionLocations.H
+ * \brief Includes the headers for all available interaction locations
+ */
+#ifndef _INTERACTION_LOCATIONS_H_
+#define _INTERACTION_LOCATIONS_H_
+
+#include "Locations.H"
+#include "InteractionLocGeneric.H"
+#include "InteractionLocHome.H"
+#include "InteractionLocWork.H"
+#include "InteractionLocSchool.H"
+
+#endif
